update code strcture
