+++ .gitignore
@@ -9,6 +9,7 @@
 /.idea/assetWizardSettings.xml
 .DS_Store
 /build
+/app/build
 .idea
 /captures
 .externalNativeBuild
