+++ app/src/main/AndroidManifest.xml
@@ -13,15 +13,15 @@
             android:supportsRtl="true"
             android:theme="@style/AppTheme">
         <activity
-                android:name=".UserHomepage"
+                android:name=".view.homepage.UserHomepage"
                 android:label="@string/title_activity_user_homepage"
                 android:theme="@style/AppTheme.NoActionBar"></activity>
         <activity
-                android:name=".RMHomepage"
+                android:name=".view.homepage.RMHomepage"
                 android:label="RMHomepage"
                 android:theme="@style/AppTheme.NoActionBar"/>
         <activity
-                android:name=".AdminHomepage"
+                android:name=".view.homepage.AdminHomepage"
                 android:label="AdminHomepage"
                 android:theme="@style/AppTheme.NoActionBar"/>
 
@@ -33,10 +33,10 @@
             </intent-filter>
         </activity>
         <activity
-                android:name=".Register"
+                android:name=".view.common.Register"
                 android:parentActivityName=".MainActivity" />
-        <activity android:name=".ChangePasswordScreen"></activity>
-        <activity android:name=".ForgotPasswordScreen" />
+        <activity android:name=".view.common.ChangePasswordScreen"></activity>
+        <activity android:name=".view.common.ForgotPasswordScreen" />
         <activity
                 android:name=".DisplayMessageActivity"
                 android:parentActivityName=".MainActivity" />
