+++ app/src/main/assets/litepal.xml
@@ -32,7 +32,7 @@
     	</list>
     -->
     <list>
-        <mapping class="com.uta.utacarrental.database.User"/>
+        <mapping class="com.uta.utacarrental.model.User"/>
     </list>
 
     <!--
