--- app/src/main/java/com/uta/utacarrental/ForgotPasswordScreen.java
@@ -1,12 +1,0 @@
-package com.uta.utacarrental;
-
-import android.os.Bundle;
-import androidx.appcompat.app.AppCompatActivity;
-
-public class ForgotPasswordScreen extends AppCompatActivity {
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_forgot_password_screen);
-    }
-}
