+++ app/src/main/java/com/uta/utacarrental/model/Reservation.java
@@ -1,0 +1,156 @@
+package com.uta.utacarrental.model;
+
+import org.litepal.annotation.Column;
+import org.litepal.crud.LitePalSupport;
+
+import java.io.Serializable;
+import java.util.Date;
+
+public class Reservation extends LitePalSupport implements Serializable {
+
+    @Column(unique = true)
+    private int reservationNumber;
+    private String firstName;
+    private String lastName;
+    private int carNumber;
+    private String carName;
+    private int carCapacity;
+    private Date startDate;
+    private Date startTime;
+    private Date endDate;
+    private Date endTime;
+    private String riderNumber;
+    private double totalCost;
+    private boolean gps;
+    private boolean sirius;
+    private boolean onStart;
+    private boolean member;
+
+    public int getReservationNumber() {
+        return reservationNumber;
+    }
+
+    public void setReservationNumber(int reservationNumber) {
+        this.reservationNumber = reservationNumber;
+    }
+
+    public String getFirstName() {
+        return firstName;
+    }
+
+    public void setFirstName(String firstName) {
+        this.firstName = firstName;
+    }
+
+    public String getLastName() {
+        return lastName;
+    }
+
+    public void setLastName(String lastName) {
+        this.lastName = lastName;
+    }
+
+    public int getCarNumber() {
+        return carNumber;
+    }
+
+    public void setCarNumber(int carNumber) {
+        this.carNumber = carNumber;
+    }
+
+    public String getCarName() {
+        return carName;
+    }
+
+    public void setCarName(String carName) {
+        this.carName = carName;
+    }
+
+    public int getCarCapacity() {
+        return carCapacity;
+    }
+
+    public void setCarCapacity(int carCapacity) {
+        this.carCapacity = carCapacity;
+    }
+
+    public Date getStartDate() {
+        return startDate;
+    }
+
+    public void setStartDate(Date startDate) {
+        this.startDate = startDate;
+    }
+
+    public Date getStartTime() {
+        return startTime;
+    }
+
+    public void setStartTime(Date startTime) {
+        this.startTime = startTime;
+    }
+
+    public Date getEndDate() {
+        return endDate;
+    }
+
+    public void setEndDate(Date endDate) {
+        this.endDate = endDate;
+    }
+
+    public Date getEndTime() {
+        return endTime;
+    }
+
+    public void setEndTime(Date endTime) {
+        this.endTime = endTime;
+    }
+
+    public String getRiderNumber() {
+        return riderNumber;
+    }
+
+    public void setRiderNumber(String riderNumber) {
+        this.riderNumber = riderNumber;
+    }
+
+    public double getTotalCost() {
+        return totalCost;
+    }
+
+    public void setTotalCost(double totalCost) {
+        this.totalCost = totalCost;
+    }
+
+    public boolean isGps() {
+        return gps;
+    }
+
+    public void setGps(boolean gps) {
+        this.gps = gps;
+    }
+
+    public boolean isSirius() {
+        return sirius;
+    }
+
+    public void setSirius(boolean sirius) {
+        this.sirius = sirius;
+    }
+
+    public boolean isOnStart() {
+        return onStart;
+    }
+
+    public void setOnStart(boolean onStart) {
+        this.onStart = onStart;
+    }
+
+    public boolean isMember() {
+        return member;
+    }
+
+    public void setMember(boolean member) {
+        this.member = member;
+    }
+}
