--- app/src/main/java/com/uta/utacarrental/ui/profile/ProfileViewModel.java
@@ -1,19 +1,0 @@
-package com.uta.utacarrental.ui.profile;
-
-import androidx.lifecycle.LiveData;
-import androidx.lifecycle.MutableLiveData;
-import androidx.lifecycle.ViewModel;
-
-public class ProfileViewModel extends ViewModel {
-
-    private MutableLiveData<String> mText;
-
-    public ProfileViewModel() {
-        mText = new MutableLiveData<>();
-        mText.setValue("This is profile fragment");
-    }
-
-    public LiveData<String> getText() {
-        return mText;
-    }
-}
