+++ app/src/main/java/com/uta/utacarrental/view/common/DisplayMessageActivity.java
@@ -1,0 +1,25 @@
+package com.uta.utacarrental.view.common;
+
+import android.content.Intent;
+import android.os.Bundle;
+import android.widget.TextView;
+import androidx.appcompat.app.AppCompatActivity;
+
+import com.uta.utacarrental.MainActivity;
+import com.uta.utacarrental.R;
+
+public class DisplayMessageActivity extends AppCompatActivity {
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_display_message);
+
+        // Get the Intent that started this activity and extract the string
+        Intent intent = getIntent();
+        String message = intent.getStringExtra(MainActivity.EXTRA_MESSAGE);
+
+        // Capture the layout's TextView and set the string as its text
+        TextView textView = findViewById(R.id.textView2);
+        textView.setText(message);
+    }
+}
