+++ app/src/main/java/com/uta/utacarrental/view/common/ForgotPasswordScreen.java
@@ -1,0 +1,14 @@
+package com.uta.utacarrental.view.common;
+
+import android.os.Bundle;
+import androidx.appcompat.app.AppCompatActivity;
+
+import com.uta.utacarrental.R;
+
+public class ForgotPasswordScreen extends AppCompatActivity {
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_forgot_password_screen);
+    }
+}
