+++ app/src/main/java/com/uta/utacarrental/view/profile/ProfileViewModel.java
@@ -1,0 +1,19 @@
+package com.uta.utacarrental.view.profile;
+
+import androidx.lifecycle.LiveData;
+import androidx.lifecycle.MutableLiveData;
+import androidx.lifecycle.ViewModel;
+
+public class ProfileViewModel extends ViewModel {
+
+    private MutableLiveData<String> mText;
+
+    public ProfileViewModel() {
+        mText = new MutableLiveData<>();
+        mText.setValue("This is profile fragment");
+    }
+
+    public LiveData<String> getText() {
+        return mText;
+    }
+}
