+++ app/src/main/java/com/uta/utacarrental/view/reservation/ReservationViewModel.java
@@ -1,0 +1,19 @@
+package com.uta.utacarrental.view.reservation;
+
+import androidx.lifecycle.LiveData;
+import androidx.lifecycle.MutableLiveData;
+import androidx.lifecycle.ViewModel;
+
+public class ReservationViewModel extends ViewModel {
+
+    private MutableLiveData<String> mText;
+
+    public ReservationViewModel() {
+        mText = new MutableLiveData<>();
+        mText.setValue("This is reservation fragment");
+    }
+
+    public LiveData<String> getText() {
+        return mText;
+    }
+}
