+++ app/src/main/java/com/uta/utacarrental/view/send/SendViewModel.java
@@ -1,0 +1,19 @@
+package com.uta.utacarrental.view.send;
+
+import androidx.lifecycle.LiveData;
+import androidx.lifecycle.MutableLiveData;
+import androidx.lifecycle.ViewModel;
+
+public class SendViewModel extends ViewModel {
+
+    private MutableLiveData<String> mText;
+
+    public SendViewModel() {
+        mText = new MutableLiveData<>();
+        mText.setValue("This is send fragment");
+    }
+
+    public LiveData<String> getText() {
+        return mText;
+    }
+}
