+++ app/src/main/res/layout/activity_forgot_password_screen.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ForgotPasswordScreen">
+    tools:context=".view.common.ForgotPasswordScreen">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
